pyproject.toml
[tool.poetry]
name = "nlp processing pipeline"
version = "0.1.0"
description = "A package to process raw disaster tweets"
authors = ["Ghazwan Namou Jablak <ghazwan.jablak@yahoo.com>"]
readme = "README.md"
packages = [
    { include = "src" },
    { include = "queries"}
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
jinja2 = "^3.1.6"
pandas = "^2.2.3"
textblob = "^0.19.0"
autocorrect = "^2.6.1"
matplotlib = "^3.10.1"
contractions = "^0.1.73"
spacy = "^3.8.4"
python-decouple = "^3.8"
sqlmodel = "0.0.8"
sqlalchemy-redshift = "^0.8.14"
psycopg2-binary = "^2.9.10"
pytest = "^8.3.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
